Add new session alert
