+++ packages/react-email-templates/emails/session-insights.tsx
@@ -1,26 +1,8 @@
-import {
-	Column,
-	Heading,
-	Hr,
-	Img,
-	Link,
-	Section,
-	Text,
-} from '@react-email/components'
+import { Heading, Hr, Img, Link, Text } from '@react-email/components'
 import * as React from 'react'
 
 import { EmailHtml, HighlightLogo } from '../components/common'
-
-interface InterestingSession {
-	identifier: string
-	avatarUrl: string
-	screenshotUrl: string
-	activityGraphUrl: string
-	country: string
-	activeLength: string
-	url: string
-	insights: string[]
-}
+import { Session, SessionPreview } from '../components/sessions'
 
 export interface SessionInsightsEmailProps {
 	projectId?: number
@@ -28,7 +10,7 @@ export interface SessionInsightsEmailProps {
 	toEmail?: string
 	unsubscribeUrl?: string
 	useHarold?: boolean
-	interestingSessions?: InterestingSession[]
+	interestingSessions?: Session[]
 }
 
 const sessionExample = {
@@ -68,72 +50,9 @@ export const SessionInsightsEmail = ({
 			Here are 3 interesting* sessions recorded in your project this week:
 		</Text>
 		<Hr style={hr} />
-		{interestingSessions.map((s, idx) => (
+		{interestingSessions.map((s) => (
 			<>
-				<Section width={378}>
-					<Link href={s.url}>
-						<Img
-							src={s.screenshotUrl}
-							style={sessionScreenshot}
-							width={378}
-							height={206}
-						/>
-					</Link>
-				</Section>
-				<Section width={378} style={sessionAttributes}>
-					<Column align="left" width={266} style={sessionProperties}>
-						<Section align="left" style={leftAlign}>
-							<Text style={identifier}>
-								{s.avatarUrl && (
-									<>
-										<Img
-											src={s.avatarUrl}
-											width="22"
-											height="22"
-											style={avatar}
-										/>
-										&nbsp;
-									</>
-								)}
-								<a style={identifierText}>{s.identifier}</a>
-							</Text>
-						</Section>
-						<Section align="left" style={leftAlign}>
-							<span style={sessionProperty}>{s.country}</span>
-							&nbsp;
-							<span style={sessionProperty}>
-								{s.activeLength}
-							</span>
-						</Section>
-						<Section
-							align="left"
-							style={{ ...leftAlign, marginTop: '8px' }}
-						>
-							<Link style={viewSessionText} href={s.url}>
-								View Session →
-							</Link>
-						</Section>
-					</Column>
-					<Column width={132} style={activityGraphColumn}>
-						<Img
-							style={activityGraph}
-							src={s.activityGraphUrl}
-							width="132"
-							height="66"
-							alt="Session activity graph"
-						/>
-					</Column>
-				</Section>
-				{s.insights.map((i, idx) => (
-					<Section width={378} key={idx} style={insight}>
-						<Column width={24} style={numberLabel}>
-							{idx + 1}
-						</Column>
-						<Column width={354} style={{ wordBreak: 'break-all' }}>
-							<Text style={insightText}>{i}</Text>
-						</Column>
-					</Section>
-				))}
+				<SessionPreview session={s} />
 				<Hr style={hr} />
 			</>
 		))}
@@ -223,99 +142,9 @@ const paragraph = {
 	color: '#ffffff',
 }
 
-const insight = {
-	border: '1px solid #30294e',
-	borderRadius: '6px',
-	padding: '8px',
-	marginTop: '8px',
-	fontSize: '14px',
-}
-
-const numberLabel = {
-	color: '#9d97aa',
-	backgroundColor: '#30294e',
-	borderRadius: '3px',
-}
-
-const sessionScreenshot = {
-	backgroundImage:
-		'url("https://static.highlight.io/assets/session-insights/session4.png")',
-	borderRadius: '6px',
-	objectFit: 'cover' as const,
-	width: '378px',
-	height: '206px',
-}
-
-const identifier = {
-	color: '#ffffff',
-	margin: '8px 0',
-	fontSize: '16px',
-	height: '24px',
-	overflow: 'hidden',
-	textAlign: 'left' as const,
-}
-
-const identifierText = {
-	color: '#ffffff',
-	textDecoration: 'none',
-}
-
-const sessionAttributes = {
-	marginBottom: '16px',
-	marginTop: '8px',
-}
-
-const viewSessionText = {
-	...anchor,
-	fontSize: '16px',
-}
-
 const subtitleText = {
 	...text,
 	fontSize: '16px',
-}
-
-const leftAlign = {
-	textAlign: 'left' as const,
-}
-
-const sessionProperties = {
-	textAlign: 'left' as const,
-	fontSize: '16px',
-}
-
-const sessionProperty = {
-	color: '#ffffff',
-	backgroundColor: 'rgba(255,255,255,0.32)',
-	borderRadius: '100px',
-	lineHeight: '26px',
-	padding: '2px 12px',
-	marginRight: '6px',
-}
-
-const avatar = {
-	display: 'inline',
-	borderRadius: '100px',
-	verticalAlign: 'middle',
-	marginRight: '4px',
-}
-
-const insightText = {
-	...text,
-	paddingLeft: '8px',
-	textAlign: 'left' as const,
-	wordBreak: 'break-word' as const,
-	margin: '8px 0',
-}
-
-const activityGraph = {
-	width: '132px',
-	height: '66px',
-}
-
-const activityGraphColumn = {
-	paddingTop: '8px',
-	verticalAlign: 'middle',
 }
 
 const headingText = {
